final commit after code enhance

+++ src/main/java/com/globel/library/service/implementation/BookServiceImpl.java
@@ -19,19 +19,22 @@ import jakarta.transaction.Transactional;
 @Service
 public class BookServiceImpl implements BookService {
 
+	// inject bookRepo
 	@Autowired
 	private BookRepo bookRepo;
 
-	// find all books
 	@Override
-	@Cacheable(value = "findAllBook",key = "#root.methodName")
+	@Cacheable(value = "findAllBook", key = "#root.methodName") // returned data into cache
 	public List<Book> findAll() {
-		return bookRepo.findAll();
+		List<Book> books = bookRepo.findAll();
+		if (books.isEmpty()) {
+			throw new RecordNotFoundException("we not have any books");
+		}
+		return books;
 	}
 
-	// find book by id
 	@Override
-	@Cacheable(value = "findBookById",key = "#id")
+	@Cacheable(value = "findBookById", key = "#id") // returned data into cache
 	public Book findById(Long id) {
 		Optional<Book> entity = bookRepo.findById(id);
 		if (entity.isPresent()) {
@@ -39,32 +42,33 @@ public class BookServiceImpl implements BookService {
 		} else {
 			throw new RecordNotFoundException("book with id:- " + id + "  not found");
 		}
-	} 
+	}
 
-	// find book and not must to retrive it via DB
 	@Override
 	public Book getById(Long id) {
 		return bookRepo.getById(id);
 	}
 
-	
-	// add new book 
 	@Override
-	@Transactional
-	@CacheEvict(value = {"findAllBook" ,"findBookById"} ,key = "#root.methodName" , allEntries = true)
+	@Transactional // forRoleBack when any ERROR
+	@CacheEvict(value = { "findAllBook", "findBookById" }, key = "#root.methodName", allEntries = true) // clear cache
+																										// (update it
+																										// with new data
+																										// )
 	public Book insert(Book entity) {
 		return bookRepo.save(entity);
 	}
 
-	// update  book 
-	
 	@Override
-	@Transactional
-//	@Transactional(propagation = Propagation.REQUIRED, isolation = Isolation.DEFAULT, timeout = 30, readOnly = false)
-	@CacheEvict(value = {"findAllBook" ,"findBookById"} ,key = "#root.methodName" , allEntries = true)
+	@Transactional // forRoleBack when any ERROR --- ? hash --> @Transactional(propagation =
+					// Propagation.REQUIRED, isolation = Isolation.DEFAULT, timeout = 30, readOnly =
+					// false)
+	@CacheEvict(value = { "findAllBook", "findBookById" }, key = "#root.methodName", allEntries = true) // clear cache
+																										// (update it
+																										// with new data
+																										// )
 	public Book update(Book entity) {
 		Book book = bookRepo.getById(entity.getId());
-
 		if (book != null) {
 			book.setTitle(entity.getTitle());
 			book.setAuthor(entity.getAuthor());
@@ -72,40 +76,50 @@ public class BookServiceImpl implements BookService {
 			book.setIsbn(entity.getIsbn());
 
 			return bookRepo.save(book);
+		} else {
+			throw new RecordNotFoundException("book  not found");
 		}
-		return book;
-
 	}
 
-	
 	// delete book by id
 	@Override
 	public void deleteById(Long id) {
-
-		bookRepo.deleteById(id);
-
-	}
-
-	
-	// spcific fun to edit book status (book found or not )
-	
-	@Override
-	public void updateAvliableFalse(Long bookId) {
-
-		Book book = bookRepo.findById(bookId).orElse(null);
-		if (book != null) {
-			book.setAvilable(false);
-			update(book);
+		Optional<Book> entity = bookRepo.findById(id);
+		if (entity.isPresent()) {
+			bookRepo.deleteById(id);
+		} else {
+			throw new RecordNotFoundException("book with id:- " + id + "  not found to deleted");
 		}
 	}
 
+	// <<< ------- help method ------>>>
+
+	// make book unavailable for borrow
+	@Override
+	public void updateAvliableFalse(Long bookId) {
+
+		Optional<Book> book = bookRepo.findById(bookId);
+		if (book.isPresent()) {
+			Book uBook = book.get();
+			uBook.setAvilable(false);
+			update(uBook);
+		}else {
+			throw new RecordNotFoundException("book with id : " + bookId +" not found");
+		}
+	}
+
+	// make book available to borrow
 	@Override
 	public void updateAvliableTrue(Long bookId) {
-		Book book = bookRepo.findById(bookId).orElse(null);
-		if (book != null) {
-			book.setId(bookId);
-			book.setAvilable(true);
-			update(book);
+		
+		Optional<Book> book = bookRepo.findById(bookId);
+		if (book.isPresent()) {
+			Book uBook = book.get();
+			uBook.setId(bookId);
+			uBook.setAvilable(true);
+			update(uBook);
+		}else {
+			throw new RecordNotFoundException("book with id : " + bookId +" not found");
 		}
 	}
